[cor] Rename Int3.h to VecInt.h
