+++ code/Core/Core.vcxproj.filters
@@ -146,7 +146,7 @@
     <ClInclude Include="Log\LogModule.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Math\Int3.h">
+    <ClInclude Include="Math\VecInt.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
